refactor proposal execution status route
